Time: 0 ms (100%), Space: 8.9 MB (2.03%) - LeetHub

+++ 0441-arranging-coins/0441-arranging-coins.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
     int arrangeCoins(int n) {
-        return floor(sqrt(2*n+0.25)-0.5);
+        return floor(sqrt((double)2*n+0.25)-0.5);
     }
 };
